警告文の部分の修正

+++ Src/Manager/Generic/InputManagerS.cpp
@@ -174,7 +174,7 @@ float InputManagerS::GetLStickDeg(InputManager::JOYPAD_NO no) const
 		return deg;
 	}
 	auto rad = std::atan2(knockSize.y, knockSize.x);
-	deg = rad * (180.0f / DX_PI_F);
+	deg = static_cast<float>(rad)* (180.0f / DX_PI_F);
 	deg += 90.0f;
 	deg = deg < 0 ? deg + 360 : deg;
 	return deg;
@@ -189,9 +189,9 @@ float InputManagerS::GetRStickDeg(InputManager::JOYPAD_NO no) const
 		return deg;
 	}
 	auto rad = std::atan2(knockSize.y, knockSize.x);
-	deg = rad * (180.0f / DX_PI_F);
+	deg = static_cast<float>(rad) * (180.0f / DX_PI_F);
 	deg += 90.0f;
-	deg = deg < 0 ? deg + 360 : deg;
+	deg = deg < 0.0f ? deg + 360.0f : deg;
 	return deg;
 }
